requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
httpx
pillow
pillow-heif
reportlab
python-multipart
